Add files via upload
In this updated version a util script has been developed. This Util file has functions for different tools that will require reading from files and outputting to files. Also, the function library file has been modified so it has access to the /bin dir  of the tool that needs to be used. This was done to counteract a technical problem.