Add files via upload
The util script now allows for standard outputting to the terminal or a text file. Reading from a text file (.txt) can be performed. Checks are in place to see the number of parameters passed by the user and also checks done to see if the txt file is passed using grep. 

The linux file patten check ( -f operator ) caused isses when domains where passed. So bespoke regular expression was implemented. 

Script tested and debugged for usage.