Add files via upload
In the function library the metadata function has been updated  to use the xxd function in order display the file in hex format or binary format. Also the user can decide if they want to see the head or full output. 

In the oneStopRecon file function call was incorrect so that was amended to the correct one so that it can function.