Add files via upload
This update version of the function library consists of an additional option for the username generation function.  The additional option is to run Sherlock  against the usernames. 
This is done by storing the 12 usernames in an array and the using a for loop to irritate around the list. 

Further enhancements consists of separating the username and email generation with 2 IF statements. The username if statement checks for  $# -eq 2 || $# -eq 3 and the email if statement check for  $# -eq 3.