Add files via upload
In the function library the text function has been amended. So check if the site has robots text file or security text file. And depending if they have file it will remove the file after it was downloaded but if there is not file downloaded the it will output a simple message. 

In the Util file a simple  file check function has been made.